+++ src/main/kotlin/flab/tickethub/auth/application/port/in/AuthCommandUseCase.kt
@@ -1,15 +1,15 @@
 package flab.tickethub.auth.application.port.`in`
 
+import flab.tickethub.auth.adaptor.`in`.request.LoginRequest
 import flab.tickethub.auth.adaptor.`in`.request.RefreshAccessTokenRequest
 import flab.tickethub.auth.domain.MemberPrincipal
 import flab.tickethub.auth.domain.TokenPair
-import flab.tickethub.auth.domain.TokenPayload
 
 interface AuthCommandUseCase {
 
-    fun logout(memberPrincipal: MemberPrincipal)
+    fun login(request: LoginRequest): TokenPair
 
-    fun updateRefreshToken(tokenPayload: TokenPayload): TokenPair
+    fun logout(memberPrincipal: MemberPrincipal)
 
     fun refreshAccessToken(request: RefreshAccessTokenRequest): String
