+++ src/main/kotlin/flab/tickethub/auth/application/service/AuthCommandService.kt
@@ -1,14 +1,15 @@
 package flab.tickethub.auth.application.service
 
+import flab.tickethub.auth.adaptor.`in`.request.LoginRequest
 import flab.tickethub.auth.adaptor.`in`.request.RefreshAccessTokenRequest
 import flab.tickethub.auth.application.port.`in`.AuthCommandUseCase
 import flab.tickethub.auth.domain.MemberPrincipal
 import flab.tickethub.auth.domain.TokenPair
-import flab.tickethub.auth.domain.TokenPayload
 import flab.tickethub.member.application.port.out.MemberQueryPort
 import flab.tickethub.support.error.ApiException
 import flab.tickethub.support.error.ErrorCode
 import flab.tickethub.support.security.TokenProvider
+import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -16,17 +17,20 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class AuthCommandService(
     private val memberQueryPort: MemberQueryPort,
+    private val passwordEncoder: PasswordEncoder,
     private val tokenProvider: TokenProvider
 ) : AuthCommandUseCase {
 
-    override fun updateRefreshToken(tokenPayload: TokenPayload): TokenPair {
-        val tokenPair = tokenProvider.generateTokenPair(tokenPayload)
-        val memberId = tokenPair.tokenPayload.id()
+    override fun login(request: LoginRequest): TokenPair {
+        val member = memberQueryPort.findByEmail(request.email)
+            ?: throw ApiException(ErrorCode.INVALID_EMAIL_OR_PASSWORD)
+        validatePassword(
+            expectedPassword = request.password,
+            actualPassword = member.password
+        )
 
-        val member = findById(memberId)
-        member.login(tokenPair.refreshToken)
-
-        return tokenPair
+        return tokenProvider.generateTokenPair(member)
+            .also { member.login(it.refreshToken) }
     }
 
     override fun refreshAccessToken(request: RefreshAccessTokenRequest): String {
@@ -41,6 +45,14 @@ class AuthCommandService(
     override fun logout(memberPrincipal: MemberPrincipal) {
         val member = findById(memberPrincipal.id())
         member.logout()
+    }
+
+    private fun validatePassword(
+        expectedPassword: String,
+        actualPassword: String
+    ) {
+        if (!passwordEncoder.matches(expectedPassword, actualPassword))
+            throw ApiException(ErrorCode.INVALID_EMAIL_OR_PASSWORD)
     }
 
     private fun findById(memberId: Long) =
